Add YouTube link support to !songrequest command
The !songrequest command now accepts YouTube links in addition to Spotify links and song titles. When a YouTube link is provided, the bot extracts and cleans the video title using yt-dlp, then uses it to search for the song on Spotify. This improves usability for users who want to request songs via YouTube links.